* cleaned up QwRunCondition and commmented/summarized possible methods   to access SVN revision info
* cleaned up several warning messages during compiling

  - warning: ------- may be used uninitialized in this function 
  - warning: comparison between signed and unsigned integer expressions
  - warning: unused variable ---------
  - warning: format '%d' expects type 'int', but argument 3 has type 'size_t'
  - warning: "/*" within comment
 

Monday, September  6 01:18:13 EDT 2010, jhlee



git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@1727 b9700a98-9937-0410-b777-f318d6eb5bff

+++ Parity/src/QwBlinder.cc
@@ -612,7 +612,7 @@ void QwBlinder::WriteTestValues(QwDatabase* db)
       string s_sql = s_sql_pre;
 
       // test_number
-      sprintf(s_number, "%d", i);
+      sprintf(s_number, "%d", (int) i);
       s_sql += string(s_number);
       s_sql += ", ";
